moved the common functions used in a python test suites to to a common python script
created helper function for displaying QA test summary
